backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-jose
python-multipart